basehangul v0.2.0 BaseHangul

This module provides BaseHangul encoding/decoding functionalities.

Visit the GitHub Page for more information.

Functions

decode(input)

Decodes BaseHangul-encoded string input.

Returns the decoded result as a string or a binary.

Examples

iex> BaseHangul.decode("낏뗐맸굉")
"Hello"
iex> BaseHangul.decode("갈걍꺅갉")
<<1, 2, 3, 4, 5>>

Notes

This function expects a UTF-8 encoded text. Passing an invalid BaseHangul-encoded string as an input will result in an undefined behavior.

decode(indev, outdev \\ Process.group_leader())

Takes BaseHangul-encoded data from indev IO device and writes decoded data to outdev IO device.

Returns :ok on success.

Examples

iex> {:ok, file} = File.open("some_file", [:read])
{:ok, #PID<0.XXX.0>}
iex> BaseHangul.decode(file, :stdio)
This is an encoded string.:ok

Notes

This function expects a UTF-8 encoded text. Passing an invalid BaseHangul-encoded string as an input will result in an undefined behavior.

encode(input)

Encodes input to a BaseHangul string. input can be a string or a binary.

Returns a BaseHangul-encoded string on success.

Examples

iex> BaseHangul.encode("Hello")
"낏뗐맸굉"
iex> BaseHangul.encode(<<1, 2, 3, 4, 5>>)
"갈걍꺅갉"

encode(indev, outdev \\ Process.group_leader())

Takes data from indev IO device and writes BaseHangul-encoded string to outdev IO device.

Returns :ok on success.

Examples

iex> {:ok, file} = File.open("some_file", [:read])
{:ok, #PID<0.XXX.0>}
iex> BaseHangul.encode(file, :stdio)
넥라넒굔덥램꼬먈눠멂불뮐궂뚠붉걍닐묩묻굡궂릅는놨궂몹듕낍
# ... more ...
놨궂몹듕낍닥듈럇낍뉴름댜둥:ok